[feat] migrate logic to viewmodel
