+++ discord-bot-helper/ContentView.swift
@@ -11,6 +11,7 @@ import Alamofire
 struct ContentView: View {
     
     @State var inputName = ""
+    private var viewModel = SendMessageViewModel()
     
     var body: some View {
         VStack {
@@ -20,41 +21,12 @@ struct ContentView: View {
             Text("Hello, world!")
             TextField("URLを入れてください", text: $inputName)
             Button(action: {
-                postDiscordWebhook(url: inputName)
+                viewModel.postDiscordWebhook(url: inputName)
             }, label: {
                 Text("Send")
             })
         }
         .padding()
-    }
-}
-
-private func postDiscordWebhook(url: String) {
-    let baseUrlString = url
-    let param: Parameters = [
-            "content": "Hello"
-        ]
-    let headers: HTTPHeaders = ["Content-Type": "application/json"]
-
-    
-    var request = URLRequest(url: (URL(string: baseUrlString) ?? URL(string: "https://www.apple.com/")!))
-    request.httpMethod = HTTPMethod.post.rawValue
-    request.setValue("application/json", forHTTPHeaderField: "Content-Type")
-    request.httpBody = NSString(
-        data: try! JSONSerialization.data(withJSONObject: param as Any, 
-                                          options:JSONSerialization.WritingOptions.prettyPrinted
-                                         ),
-        encoding: String.Encoding.utf8.rawValue
-    )!
-        .data(using: String.Encoding.utf8.rawValue)
-    AF.request(request)
-    .responseData { response in
-        switch response.result {
-            case .success(let data):
-                print("success")
-            case .failure(let error):
-                print("error")
-        }
     }
 }
