+++ discord-bot-helper/SendMessageViewModel.swift
@@ -6,3 +6,36 @@
 //
 
 import Foundation
+import Alamofire
+
+struct SendMessageViewModel {
+    public func postDiscordWebhook(url: String) {
+        let baseUrlString = url
+        let param: Parameters = [
+                "content": "Hello"
+            ]
+        let headers: HTTPHeaders = ["Content-Type": "application/json"]
+
+        
+        var request = URLRequest(url: (URL(string: baseUrlString) ?? URL(string: "https://www.apple.com/")!))
+        request.httpMethod = HTTPMethod.post.rawValue
+        request.setValue("application/json", forHTTPHeaderField: "Content-Type")
+        request.httpBody = NSString(
+            data: try! JSONSerialization.data(withJSONObject: param as Any,
+                                              options:JSONSerialization.WritingOptions.prettyPrinted
+                                             ),
+            encoding: String.Encoding.utf8.rawValue
+        )!
+            .data(using: String.Encoding.utf8.rawValue)
+        print(request)
+        AF.request(request)
+        .responseData { response in
+            switch response.result {
+                case .success(let data):
+                    print("success")
+                case .failure(let error):
+                    print("error")
+            }
+        }
+    }
+}
